Minimal additions to workaround SIP
macOS's SIP system results in programs in the protected areas running
in a stripped environment to minimize security issues. In particular,
this results in the dynamic loader path, DYLD_LIBRARY_PATH being
emptied, but this is required by MessageFacility, Art etc to load
plugins and I/O dictionaries.

Whilst a longer term solution is worked on, provide a small set
of workarounds to that allow build time programs/tests to be run
is programs outside the protected area. The primary change is
to replace direct shbangs by #!/usr/bin/env <name>.

When additional arguments are supplied to to perl in scripts,
workaround by using `use` statements.